Fixed what should be the last of the newely induced panning bugs which was in this instance caused by a cast to too small a type

+++ libAudio/moduleMixer/moduleMixer.cpp
@@ -656,7 +656,7 @@ inline void ModuleFile::PanningSlide(Channel *channel, uint8_t param)
 	{
 		panningSlide += channel->RawPanning;
 		clipInt<int16_t>(panningSlide, 0, 256);
-		channel->RawPanning = uint8_t(panningSlide);
+		channel->RawPanning = uint16_t(panningSlide);
 	}
 }
